+++ base_layer/core/src/base_node/comms_interface/error.rs
@@ -22,7 +22,7 @@
 
 use tari_common_types::types::FixedHash;
 use tari_comms_dht::outbound::DhtOutboundError;
-use tari_node_components::blocks::BlockHeaderValidationError;
+use tari_node_components::blocks::{BlockError, BlockHeaderValidationError};
 use tari_service_framework::reply_channel::TransportChannelError;
 use tari_transaction_components::{
     consensus::ConsensusManagerError,
@@ -34,7 +34,6 @@ use tari_transaction_components::{
 use thiserror::Error;
 
 use crate::{
-    blocks::BlockError,
     chain_storage::ChainStorageError,
     mempool::MempoolError,
     proof_of_work::{cuckaroo_pow::CuckarooVerificationError, monero_rx::MergeMineError},
