+++ base_layer/core/src/base_node/comms_interface/local_interface.rs
@@ -27,7 +27,7 @@ use tari_common_types::{
     epoch::VnEpoch,
     types::{BlockHash, CompressedCommitment, CompressedPublicKey, CompressedSignature, FixedHash, HashOutput},
 };
-use tari_node_components::blocks::{Block, NewBlockTemplate};
+use tari_node_components::blocks::{Block, ChainHeader, HistoricalBlock, NewBlockTemplate};
 use tari_service_framework::{reply_channel::SenderService, Service};
 use tari_transaction_components::{
     tari_proof_of_work::{Difficulty, PowAlgorithm},
@@ -44,7 +44,6 @@ use crate::{
         NodeCommsRequest,
         NodeCommsResponse,
     },
-    blocks::{ChainHeader, HistoricalBlock},
     chain_storage::{
         InputMinedInfo,
         MinedInfo,
