+++ base_layer/core/src/base_node/sync/horizon_state_sync/synchronizer.rs
@@ -32,7 +32,7 @@ use log::*;
 use tari_common_types::types::{CompressedCommitment, FixedHash, RangeProofService};
 use tari_comms::{connectivity::ConnectivityRequester, peer_manager::NodeId, protocol::rpc::RpcClient, PeerConnection};
 use tari_crypto::commitment::HomomorphicCommitment;
-use tari_node_components::blocks::BlockHeader;
+use tari_node_components::blocks::{BlockHeader, ChainHeader};
 use tari_transaction_components::{
     transaction_components::{transaction_output::batch_verify_range_proofs, TransactionKernel, TransactionOutput},
     validation::{aggregate_body::validate_individual_output, helpers::validate_output_version},
@@ -52,7 +52,7 @@ use crate::{
         BlockchainSyncConfig,
         SyncPeer,
     },
-    blocks::{ChainHeader, UpdateBlockAccumulatedData},
+    blocks::UpdateBlockAccumulatedData,
     chain_storage::{async_db::AsyncBlockchainDb, BlockchainBackend, ChainStorageError, MmrTree},
     common::rolling_avg::RollingAverageTime,
     consensus::BaseNodeConsensusManager,
