+++ base_layer/core/src/chain_storage/db_transaction.rs
@@ -28,12 +28,12 @@ use std::{
 
 use primitive_types::U512;
 use tari_common_types::types::{BlockHash, CompressedCommitment, HashOutput};
-use tari_node_components::blocks::{Block, BlockHeader};
+use tari_node_components::blocks::{Block, BlockHeader, BlockHeaderAccumulatedData, ChainBlock, ChainHeader};
 use tari_transaction_components::transaction_components::{OutputType, TransactionKernel, TransactionOutput};
 use tari_utilities::hex::Hex;
 
 use crate::{
-    blocks::{BlockHeaderAccumulatedData, ChainBlock, ChainHeader, UpdateBlockAccumulatedData},
+    blocks::UpdateBlockAccumulatedData,
     chain_storage::{error::ChainStorageError, HorizonData, Reorg},
 };
 #[derive(Debug)]
