+++ base_layer/core/src/chain_storage/error.rs
@@ -23,6 +23,7 @@
 use lmdb_zero::error;
 use tari_common_types::{chain_metadata::ChainMetaDataError, types::FixedHashSizeError};
 use tari_mmr::{error::MerkleMountainRangeError, sparse_merkle_tree::SMTError, MerkleProofError};
+use tari_node_components::blocks::BlockError;
 use tari_storage::lmdb_store::LMDBError;
 use tari_transaction_components::{
     tari_proof_of_work::PowError,
@@ -34,7 +35,7 @@ use tari_utilities::ByteArrayError;
 use thiserror::Error;
 use tokio::task;
 
-use crate::{blocks::BlockError, chain_storage::MmrTree, validation::ValidationError, MrHashError};
+use crate::{chain_storage::MmrTree, validation::ValidationError, MrHashError};
 #[derive(Debug, Error)]
 pub enum ChainStorageError {
     #[error("Access to the underlying storage mechanism failed: {0}")]
