+++ base_layer/core/src/proto/block.rs
@@ -23,11 +23,10 @@
 use std::convert::{TryFrom, TryInto};
 
 use tari_common_types::types::PrivateKey;
-use tari_node_components::blocks::{Block, NewBlock};
+use tari_node_components::blocks::{Block, HistoricalBlock, NewBlock};
 use tari_utilities::ByteArray;
 
 use super::core as proto;
-use crate::blocks::HistoricalBlock;
 //---------------------------------- Block --------------------------------------------//
 
 impl TryFrom<proto::Block> for Block {
