+++ base_layer/core/src/validation/mocks.rs
@@ -29,7 +29,7 @@ use tari_common_types::{
     chain_metadata::ChainMetadata,
     types::{CompressedCommitment, FixedHash},
 };
-use tari_node_components::blocks::{Block, BlockHeader};
+use tari_node_components::blocks::{Block, BlockHeader, ChainBlock};
 use tari_transaction_components::{tari_proof_of_work::Difficulty, transaction_components::Transaction};
 use tari_utilities::epoch_time::EpochTime;
 
@@ -41,7 +41,6 @@ use super::{
     TransactionValidator,
 };
 use crate::{
-    blocks::ChainBlock,
     chain_storage::BlockchainBackend,
     proof_of_work::{randomx_factory::RandomXFactory, AchievedTargetDifficulty},
     test_helpers::create_consensus_rules,
