+++ base_layer/core/src/validation/traits.rs
@@ -24,12 +24,11 @@ use tari_common_types::{
     chain_metadata::ChainMetadata,
     types::{CompressedCommitment, FixedHash},
 };
-use tari_node_components::blocks::{Block, BlockHeader};
+use tari_node_components::blocks::{Block, BlockHeader, ChainBlock};
 use tari_transaction_components::{tari_proof_of_work::Difficulty, transaction_components::Transaction};
 use tari_utilities::epoch_time::EpochTime;
 
 use crate::{
-    blocks::ChainBlock,
     chain_storage::BlockchainBackend,
     proof_of_work::AchievedTargetDifficulty,
     validation::error::ValidationError,
