+++ base_layer/core/tests/tests/node_service.rs
@@ -31,7 +31,6 @@ use tari_core::{
         comms_interface::BlockEvent,
         state_machine_service::states::{events_and_states::ListeningInfo, StateInfo, StatusInfo},
     },
-    blocks::ChainBlock,
     chain_storage::BlockchainDatabaseConfig,
     consensus::{BaseNodeConsensusManager, BaseNodeConsensusManagerBuilder},
     mempool::TxStorageResponse,
@@ -43,7 +42,7 @@ use tari_core::{
         DifficultyCalculator,
     },
 };
-use tari_node_components::blocks::NewBlock;
+use tari_node_components::blocks::{ChainBlock, NewBlock};
 use tari_test_utils::unpack_enum;
 use tari_transaction_components::{
     consensus::NetworkConsensus,
