+++ base_layer/node_components/Cargo.toml
@@ -14,6 +14,7 @@ wasm = ["wasm-bindgen"]
 [dependencies]
 tari_common_types = { workspace = true }
 tari_crypto = { workspace = true, features = ["borsh"] }
+tari_hashing = { workspace = true }
 tari_transaction_components = { workspace = true }
 tari_utilities = { workspace = true, features = ["borsh"] }
 
@@ -22,6 +23,7 @@ borsh = { version = "1.5.7", features = ["derive"] }
 chrono = { version = "0.4", default-features = false, features = ["serde"] }
 digest = "0.10"
 log = "0.4"
+primitive-types = { version = "0.12", features = ["serde"] }
 serde = { version = "1.0.106", features = ["derive"] }
 thiserror = "1.0.26"
 wasm-bindgen = { version = "0.2.100", optional = true }
