+++ base_layer/node_components/src/blocks/block_header.rs
@@ -49,6 +49,7 @@ use chrono::{DateTime, Utc};
 use digest::consts::U32;
 use serde::{Deserialize, Serialize};
 use tari_common_types::types::{BlockHash, FixedHash, PrivateKey};
+use tari_hashing::BlocksHashDomain;
 use tari_transaction_components::{
     consensus::DomainSeparatedConsensusHasher,
     tari_proof_of_work::{PowAlgorithm, PowError, ProofOfWork},
@@ -56,7 +57,7 @@ use tari_transaction_components::{
 use tari_utilities::{epoch_time::EpochTime, hex::Hex};
 use thiserror::Error;
 
-use crate::blocks::{BlockBuilder, BlocksHashDomain, NewBlockHeaderTemplate};
+use crate::blocks::{BlockBuilder, NewBlockHeaderTemplate};
 #[derive(Debug, Error)]
 pub enum BlockHeaderValidationError {
     #[error("The Genesis block header is incorrectly chained")]
