+++ base_layer/node_components/src/blocks/historical_block.rs
@@ -23,9 +23,8 @@
 use std::{fmt, fmt::Display, sync::Arc};
 
 use tari_common_types::types::HashOutput;
-use tari_node_components::blocks::{Block, BlockHeader};
 
-use crate::blocks::{error::BlockError, BlockHeaderAccumulatedData, ChainBlock};
+use crate::blocks::{error::BlockError, Block, BlockHeader, BlockHeaderAccumulatedData, ChainBlock};
 
 /// The representation of a historical block in the blockchain. It is essentially identical to a protocol-defined
 /// block but contains some extra metadata that clients such as Block Explorers will find interesting.
