+++ base_layer/transaction_components/src/tari_proof_of_work/accumulated_difficulty.rs
@@ -23,8 +23,9 @@
 use std::fmt;
 
 use serde::{Deserialize, Serialize};
-use tari_transaction_components::tari_proof_of_work::{Difficulty, DifficultyError, MIN_DIFFICULTY};
 use tari_utilities::ByteArray;
+
+use crate::tari_proof_of_work::{Difficulty, DifficultyError, MIN_DIFFICULTY};
 /// The difficulty is defined as the maximum target divided by the block hash.
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, Ord, Deserialize, Serialize)]
 pub struct AccumulatedDifficulty(u128);
